ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
pandas
scikit-learn
joblib
requests
python-jose
